refactor(llm): fold Anthropic version gates into model_capabilities
model_capabilities.py (added on main in 5a2f8f4a81) is the natural home for
these name-parsed capability checks: it is the module for per-model capability
lookups and is deliberately lightweight to import (no DB deps), which was the
only reason anthropic_versions.py existed as a separate file.
